Further refactoring for documentation
