+++ src/Resources/ModelRestResource.php
@@ -1,0 +1,601 @@
+<?php
+
+/*
+ *	Copyright 2015 RhubarbPHP
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+namespace Rhubarb\RestApi\Resources;
+
+require_once __DIR__ . '/CollectionRestResource.php';
+
+use Rhubarb\Crown\DateTime\RhubarbDateTime;
+use Rhubarb\Crown\Logging\Log;
+use Rhubarb\RestApi\Exceptions\InsertException;
+use Rhubarb\RestApi\Exceptions\RestImplementationException;
+use Rhubarb\RestApi\Exceptions\UpdateException;
+use Rhubarb\RestApi\UrlHandlers\RestHandler;
+use Rhubarb\Stem\Collections\Collection;
+use Rhubarb\Stem\Exceptions\RecordNotFoundException;
+use Rhubarb\Stem\Filters\Equals;
+use Rhubarb\Stem\Models\Model;
+use Rhubarb\Stem\Schema\Relationships\ManyToMany;
+use Rhubarb\Stem\Schema\Relationships\OneToMany;
+use Rhubarb\Stem\Schema\SolutionSchema;
+
+/**
+ * A RestResource that understands how to handle collections and items when linked to Stem model.
+ */
+abstract class ModelRestResource extends CollectionRestResource
+{
+    /**
+     * A list of model names to model resource class names.
+     *
+     * @var string[]
+     */
+    public static $modelToResourceMapping = [];
+
+    protected $model = false;
+
+    public function __construct(RestResource $parentResource = null)
+    {
+        parent::__construct($parentResource);
+    }
+
+    /**
+     * Turns a model into a flat array structure ready for returning as a resource response.
+     *
+     * @param string[] $columns The columns to extract
+     * @return array A key/value pairing of columns and values
+     * @throws RestImplementationException
+     */
+    protected function transformModelToArray($columns)
+    {
+        $model = $this->getModel();
+
+        $extract = [];
+
+        $relationships = null;
+
+        foreach ($columns as $label => $column) {
+            $apiLabel = (is_numeric($label)) ? $column : $label;
+
+            $value = $model->$column;
+
+            if ( is_object( $value ) ) {
+                // We can't pass objects back through the API! Let's get a JSON friendly structure instead.
+                if (!($value instanceof Model) && !($value instanceof Collection)) {
+                    // This seems strange however if we just used json_encode we'd be passing the encoded version
+                    // back as a string. We decode to get the original structure back again.
+                    $value = json_decode(json_encode($value));
+                } else {
+                    // Look for resource modifiers after the column name
+                    $modifier = "";
+                    $urlSuffix = false;
+                    $relatedField = false;
+
+                    if (stripos($column, ":") !== false) {
+                        $parts = explode(":", $column);
+                        $column = $parts[0];
+
+                        if (is_numeric($label)) {
+                            $apiLabel = $column;
+                        }
+
+                        $modifier = strtolower($parts[1]);
+
+                        if (sizeof($parts) > 2) {
+                            $urlSuffix = $parts[2];
+                        }
+                    } else {
+                        if (stripos($column, ".") !== false) {
+                            $parts = explode(".", $column, 2);
+                            $column = $parts[0];
+                            $relatedField = $parts[1];
+
+                            if (is_numeric($label)) {
+                                $apiLabel = $parts[1];
+                            }
+                        }
+                    }
+
+                    $navigationValue = $value;
+                    $navigationResource = false;
+
+                    if ($navigationValue instanceof Model) {
+                        if ($relatedField) {
+                            eval('$extract[ $apiLabel ] = $navigationValue->' . str_replace(".", "->",
+                                    $relatedField) . ';');
+                            continue;
+                        }
+
+                        $navigationResource = $this->getRestResourceForModel($navigationValue);
+
+                        if ($navigationResource === false) {
+                            throw new RestImplementationException(print_r($navigationValue, true));
+                            continue;
+                        }
+                    }
+
+                    if ($navigationValue instanceof Collection) {
+                        $navigationResource = $this->getRestResourceForModelName(SolutionSchema::getModelNameFromClass($navigationValue->getModelClassName()));
+
+                        if ($navigationResource === false) {
+                            continue;
+                        }
+
+                        $navigationResource = $navigationResource->getCollection();
+                        $navigationResource->setModelCollection($navigationValue);
+                    }
+
+                    if ($navigationResource) {
+                        switch ($modifier) {
+                            case "summary":
+                                $value = $navigationResource->summary();
+                                break;
+                            case "link":
+                                $link = $navigationResource->link();
+
+                                if ($urlSuffix != "") {
+                                    $ourHref = $this->getRelativeUrl($_SERVER["SCRIPT_NAME"]);
+
+                                    // Override the href with this appendage instead.
+                                    $link->href = $ourHref . $urlSuffix;
+                                }
+
+                                $value = $link;
+
+                                break;
+                            default:
+                                $value = $navigationResource->get();
+                                break;
+                        }
+                    }
+                }
+            }
+
+            if ( $value !== null ) {
+                $extract[$apiLabel] = $value;
+            }
+        }
+
+        return $extract;
+    }
+
+    /**
+     * Override to control the columns returned in HEAD requests
+     *
+     * @return string[]
+     */
+    protected function getSummaryColumns()
+    {
+        $columns = [];
+
+        $model = $this->getSampleModel();
+        $columnName = $model->getLabelColumnName();
+
+        if ( $columnName != "" ){
+            $columns[] = $columnName;
+        }
+
+        return $columns;
+    }
+
+    /**
+     * Override to control the columns returned in GET requests
+     *
+     * @return string[]
+     */
+    protected function getColumns()
+    {
+        return $this->getSummaryColumns();
+    }
+
+    public function summary()
+    {
+        $resource = parent::get();
+
+        $data = $this->transformModelToArray($this->getSummaryColumns());
+
+        foreach ($data as $key => $value) {
+            $resource->$key = $value;
+        }
+
+        return $resource;
+    }
+
+    public function get()
+    {
+        if ( !$this->model )
+        {
+            return parent::get();
+        }
+
+        $resource = $this->getSkeleton();
+
+        $data = $this->transformModelToArray($this->getColumns());
+
+        foreach ($data as $key => $value) {
+            $resource->$key = $value;
+        }
+
+        return $resource;
+    }
+
+    public function head()
+    {
+        if ( !$this->model )
+        {
+            return parent::get();
+        }
+
+        $resource = $this->getSkeleton();
+
+        $data = $this->transformModelToArray($this->getSummaryColumns());
+
+        foreach ($data as $key => $value) {
+            $resource->resource->$key = $value;
+        }
+
+        return $resource;
+    }
+
+    /**
+     * get's the Model object used to populate the REST resource
+     *
+     * This is public as it is sometimes required by child handlers to check security etc.
+     *
+     * @throws \Rhubarb\RestApi\Exceptions\RestImplementationException
+     * @return Model|null
+     */
+    public function getModel()
+    {
+        if ($this->model === false) {
+            $this->model = SolutionSchema::getModel($this->getModelName(), $this->id);
+        }
+
+        if (!$this->model) {
+            throw new RestImplementationException("There is no matching resource for this url");
+        }
+
+        return $this->model;
+    }
+
+    /**
+     * Sets the model that should be used for the operations of this resource.
+     *
+     * This is normally only used by collections for efficiency (to avoid constructing additional objects)
+     *
+     * @param Model $model
+     */
+    public function setModel(Model $model)
+    {
+        $this->model = $model;
+    }
+
+    /**
+     * Override to response to the event of a model being updated through a PUT
+     *
+     * @param $model
+     * @param $restResource
+     */
+    protected function beforeModelUpdated($model, $restResource)
+    {
+
+    }
+
+    protected function getSkeleton()
+    {
+        $skeleton = parent::getSkeleton();
+
+        if ( $this->model ){
+            $skeleton->_id = $this->model->UniqueIdentifier;
+        }
+
+        return $skeleton;
+    }
+
+    public function put($restResource)
+    {
+        try {
+            $model = $this->getModel();
+            $model->importData($restResource);
+
+            $this->beforeModelUpdated($model, $restResource);
+
+            $model->save();
+
+            return true;
+        } catch (RecordNotFoundException $er) {
+            throw new UpdateException("That record could not be found.");
+        } catch (\Exception $er) {
+            throw new UpdateException($er->getMessage());
+        }
+    }
+
+    public function delete()
+    {
+        try {
+            $model = $this->getModel();
+            $model->delete();
+
+            return true;
+        } catch (\Exception $er) {
+            return false;
+        }
+    }
+
+    public static function registerModelToResourceMapping($modelName, $resourceClassName)
+    {
+        self::$modelToResourceMapping[$modelName] = $resourceClassName;
+    }
+
+    public function getRestResourceForModel(Model $model)
+    {
+        $modelName = $model->getModelName();
+
+        if (!isset(self::$modelToResourceMapping[$modelName])) {
+            throw new RestImplementationException("The model $modelName does not have an associated rest resource.");
+        }
+
+        $class = self::$modelToResourceMapping[$modelName];
+
+        /**
+         * @var RestResource $resource
+         */
+        $resource = new $class();
+        $resource->setUrlHandler($this->urlHandler);
+
+        if ( $resource instanceof ModelRestResource ){
+            $resource = $resource->getItemResourceForModel($model);
+        }
+
+        return $resource;
+    }
+
+    public static function getRestResourceForModelName($modelName)
+    {
+        if (!isset(self::$modelToResourceMapping[$modelName])) {
+            return false;
+        }
+
+        $class = self::$modelToResourceMapping[$modelName];
+
+        $resource = new $class();
+
+        return $resource;
+    }
+
+    public static function clearRestResourceMapping()
+    {
+        self::$modelToResourceMapping = [];
+    }
+
+    protected function getSampleModel()
+    {
+        return SolutionSchema::getModel( $this->getModelName() );
+    }
+
+    /**
+     * @return \Rhubarb\Stem\Collections\Collection|null
+     */
+    public function getModelCollection()
+    {
+        $collection = $this->createModelCollection();
+
+        $this->filterModelCollectionForSecurity($collection);
+
+        return $collection;
+    }
+
+
+    /**
+     * Override to filter a model collection to apply any necessary filters only when this is the specific resource being fetched
+     *
+     * The default handling applies the same filters as filterModelCollectionContainer, so don't call the parent implementation unless you want that.
+     *
+     * @param Collection $collection
+     */
+    public function filterModelCollection(Collection $collection)
+    {
+        $this->filterModelCollectionContainer($collection);
+    }
+
+    /**
+     * Override to filter a model collection to apply any necessary filters only when this is a REST parent of the specific resource being fetched
+     *
+     * @param Collection $collection
+     */
+    public function filterModelCollectionContainer(Collection $collection)
+    {
+    }
+
+    public function filterModelCollectionForModifiedSince(Collection $collection, RhubarbDateTime $since)
+    {
+        throw new RestImplementationException("A collection filtered by modified date was requested however this resource does not support it.");
+    }
+
+    /**
+     * Override to filter a model collection generated by a ModelRestCollection
+     *
+     * Normally used by root collections to filter based on authentication permissions.
+     *
+     * @param Collection $collection
+     */
+    public function filterModelCollectionForSecurity(Collection $collection)
+    {
+
+    }
+
+
+    /**
+     * Returns the name of the model to use for this resource.
+     *
+     * @return string
+     */
+    public abstract function getModelName();
+
+    protected function createModelCollection()
+    {
+        // If we have a parent resource we will look to see if we can exploit a relationship
+        // to use as our starting collection. This will ensure we only serve the correct
+        // resources
+        if ($this->parentResource && ($this->parentResource instanceof ModelRestResource)) {
+            // See there is a relationship between these two models that can be exploited
+            $parentModelName = $this->parentResource->getModelName();
+            $relationships = SolutionSchema::getAllRelationshipsForModel($parentModelName);
+
+            // Our model name
+            $modelName = $this->getModelName();
+
+            foreach ($relationships as $relationship) {
+                if ($relationship instanceof ManyToMany) {
+                    if ($relationship->getRightModelName() == $modelName) {
+                        return $relationship->fetchFor($this->parentResource->getModel());
+                    }
+                }
+
+                if ($relationship instanceof OneToMany) {
+                    if ($relationship->getTargetModelName() == $modelName) {
+                        return $relationship->fetchFor($this->parentResource->getModel());
+                    }
+                }
+            }
+        }
+
+        return new Collection($this->getModelName());
+    }
+
+    public function containsResourceIdentifier($resourceIdentifier)
+    {
+        $collection = clone $this->getModelCollection();
+
+        $this->filterModelCollectionContainer($collection);
+
+        $collection->filter(new Equals($collection->getModelSchema()->uniqueIdentifierColumnName, $resourceIdentifier));
+
+        if (count($collection) > 0) {
+            return true;
+        }
+
+        return false;
+    }
+
+    protected function summarizeItems($from, $to, RhubarbDateTime $since = null)
+    {
+        return $this->fetchItems($from, $to, $since, true);
+    }
+
+    protected function getItems($from, $to, RhubarbDateTime $since = null)
+    {
+        return $this->fetchItems($from, $to, $since);
+    }
+
+    private function fetchItems($from, $to, RhubarbDateTime $since = null, $asSummary = false)
+    {
+        $collection = $this->getModelCollection();
+
+        Log::performance("Filtering collection", "RESTAPI");
+
+        $this->filterModelCollection($collection);
+
+        if ($since !== null) {
+            $this->filterModelCollectionForModifiedSince($collection, $since);
+        }
+
+        $pageSize = ($to - $from) + 1;
+        $collection->setRange($from, $pageSize);
+
+        $items = [];
+
+        Log::performance("Starting collection iteration", "RESTAPI");
+
+        foreach ($collection as $model) {
+            $resource = $this->getItemResourceForModel( $model );
+
+            $modelStructure = ($asSummary) ? $resource->summary() : $resource->get();
+            $items[] = $modelStructure;
+        }
+
+        return [$items, sizeof($collection)];
+    }
+
+    private function getItemResourceForModel( $model )
+    {
+        $resource = clone $this;
+        $resource->parentResource = $this;
+        $resource->setModel($model);
+
+        return $resource;
+    }
+
+    protected function getHref()
+    {
+        if ( $this->urlHandler != null ) {
+            $href = parent::getHref();
+
+            if ($this->model) {
+                $href .= "/" . $this->model->UniqueIdentifier;
+            }
+
+            return $href;
+        }
+    }
+
+    public function post($restResource)
+    {
+        try {
+            $newModel = SolutionSchema::getModel($this->getModelName());
+
+            if (is_array($restResource)) {
+                $newModel->importData($restResource);
+            }
+
+            $newModel->save();
+
+            $this->afterModelCreated($newModel, $restResource);
+
+            return $this->getItemResourceForModel($newModel)->get();
+        } catch (RecordNotFoundException $er) {
+            throw new InsertException("That record could not be found.");
+        } catch (\Exception $er) {
+            throw new InsertException($er->getMessage());
+        }
+    }
+
+    /**
+     * Override to respond to the event of a new model being created through a POST
+     *
+     * @param $model
+     * @param $restResource
+     */
+    protected function afterModelCreated($model, $restResource)
+    {
+
+    }
+
+    /**
+     * Returns the ItemRestResource for the $resourceIdentifier contained in this collection.
+     *
+     * @param $resourceIdentifier
+     * @return ItemRestResource
+     * @throws RestImplementationException Thrown if the item could not be found.
+     */
+    public function createItemResource($resourceIdentifier)
+    {
+        $model = SolutionSchema::getModel($this->getModelName(),$resourceIdentifier);
+
+        return $this->getItemResourceForModel($model);
+    }
+}
