+++ tests/Resources/ModelRestResourceTest.php
@@ -25,7 +25,7 @@ use Rhubarb\Crown\Request\WebRequest;
 use Rhubarb\Crown\Tests\RhubarbTestCase;
 use Rhubarb\RestApi\Authentication\AuthenticationProvider;
 use Rhubarb\RestApi\Resources\ItemRestResource;
-use Rhubarb\RestApi\Resources\ModelCollectionRestResource;
+use Rhubarb\RestApi\Resources\ModelRestResource;
 use Rhubarb\RestApi\UrlHandlers\RestApiRootHandler;
 use Rhubarb\RestApi\UrlHandlers\RestCollectionHandler;
 use Rhubarb\Stem\Schema\SolutionSchema;
@@ -242,7 +242,7 @@ class ModelRestResourceTest extends RhubarbTestCase
 
     public function testCustomColumns()
     {
-        ModelCollectionRestResource::registerModelToResourceMapping("Company", UnitTestCompanyRestResource::class);
+        ModelRestResource::registerModelToResourceMapping("Company", UnitTestCompanyRestResource::class);
 
         $context = new Context();
 
@@ -269,8 +269,8 @@ class ModelRestResourceTest extends RhubarbTestCase
 
     public function testHeadLinks()
     {
-        ModelCollectionRestResource::registerModelToResourceMapping("Company", UnitTestCompanyRestResource::class);
-        ModelCollectionRestResource::registerModelToResourceMapping("Example",
+        ModelRestResource::registerModelToResourceMapping("Company", UnitTestCompanyRestResource::class);
+        ModelRestResource::registerModelToResourceMapping("Example",
             UnitTestExampleRestResourceWithCompanyHeader::class);
 
         $context = new Context();
@@ -332,7 +332,7 @@ class ModelRestResourceTest extends RhubarbTestCase
         $response = $api->GenerateResponse($request);
         $content = $response->GetContent();
 
-        $this->assertEquals("http://cli/contacts/1", $content->_href);
+        $this->assertEquals("/contacts/1", $content->_href);
     }
 
     public function testCollectionIsFiltered()
@@ -387,7 +387,7 @@ class UnitTestDummyResource extends ItemRestResource
 
 }
 
-class UnitTestExampleRestResourceCustomisedColumns extends ModelCollectionRestResource
+class UnitTestExampleRestResourceCustomisedColumns extends ModelRestResource
 {
     protected function getColumns()
     {
@@ -406,7 +406,7 @@ class UnitTestExampleRestResourceCustomisedColumns extends ModelCollectionRestRe
     }
 }
 
-class UnitTestExampleRestResourceWithCompanyHeader extends ModelCollectionRestResource
+class UnitTestExampleRestResourceWithCompanyHeader extends ModelRestResource
 {
     protected function getColumns()
     {
@@ -424,7 +424,7 @@ class UnitTestExampleRestResourceWithCompanyHeader extends ModelCollectionRestRe
     }
 }
 
-class UnitTestExampleRestResource extends ModelCollectionRestResource
+class UnitTestExampleRestResource extends ModelRestResource
 {
     /**
      * Returns the name of the model to use for this resource.
@@ -439,13 +439,16 @@ class UnitTestExampleRestResource extends ModelCollectionRestResource
     protected function getColumns()
     {
         $columns = parent::getColumns();
+        $columns[] = "Surname";
         $columns[] = "Company";
 
         return $columns;
     }
+
+
 }
 
-class UnitTestCompanyRestResource extends ModelCollectionRestResource
+class UnitTestCompanyRestResource extends ModelRestResource
 {
     protected function getColumns()
     {
